Fix broken content_view proxying
Previously we were overwriting the :cm method on the parent class
of the template so we could intercept the call and call the correct
view_content object. Unfortunately this is permanent for the life
of the parent class. To get around this we build a new class
and proxy :method_missing calls through to the original view